move list new property component to components folder
